% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRADMM.R
\name{QRADMM}
\alias{QRADMM}
\title{Penalized quantile regression with ADMM}
\usage{
QRADMM(X, y, tau, rho, lambda, iter = 100, intercept = TRUE,
  penalty = "none", a = 3.7, lambda1 = 1, lambda2 = 1)
}
\arguments{
\item{X}{The design matrix (without intercept)}

\item{y}{The response vector}

\item{tau}{The quantile of interest}

\item{rho}{The augmentation parameter for the ADMM}

\item{lambda}{The penalization parameter}

\item{iter}{Maximum number of iterations allowed}

\item{intercept}{Whether to include the intercept into the model, default is TRUE}

\item{penalty}{Name of the penalty to use, currently support ("none","lasso","enet","scad","mcp")}

\item{a}{The shape parameter for the MCP/SCAD penalty}

\item{lambda1}{Extra penalization parameter for the elastic net. Specifically, the l1 part of the elastic net has penalization coefficient lambda*labda1}

\item{lambda2}{Extra penalization parameter for the elastic net. Specifically, the l2 part of the elastic net has penalization coefficient lambda*labda2}
}
\value{
The coefficient estimation of the linear quantile regression model
}
\description{
This is a non-parallel implementation. Currently support unpenalized quantile regression and
penalized quantile regression with the Lasso, elastic net, MCP and SCAD penalties. See the paper
\url{https://onlinelibrary.wiley.com/doi/abs/10.1111/insr.12221} for details.
}
\examples{
require(MASS)
n=300;p=1000
Sig = matrix(0, nrow = p, ncol = p)
for(i1 in 1:p){
  for(j in 1:p){
    Sig[i1,j] = 0.5^abs(i1-j)
  }
}
X = mvrnorm(n, mu = rep(0,p), Sigma = Sig, tol = 1e-6)
X[,1] = pnorm(X[,1])
u = rnorm(n)
y=X[,3] + X[,4] + X[,5] + X[,1]*u
beta = QRADMM(X, y, .9, 1, 20, penalty="scad", intercept=FALSE)
}
\seealso{
\code{\link{QPADM}}
}
